Fixed clipped group scrollbar behavior
Previously if a group had a scrollbar and is not completly visible
inside the window then the its scrollbar would use the clipped
height instead of its actual height as scroll reference. This should
now be fixed and work correctly.

In addition I split the tab/node button style into two one for the
minimized and one for the maximized state change.