Remove redundant static assertions
They're already in the API part of the header, I don't see the need
to repeat them in the implementation part of the header, as they're
exactly the same.

The header would work the same unless someone explicity by-passed
the API by defining NK_NUKLEAR_H_ before including the header, but
I just can't see why someone would do that.