BREAKING CHANGE! Changed `nk_clear` behavior #656
Previously `nk_clear` with optional vertex output enabled would try and
clear each passed buffer from `nk_convert` if called. Problems arise
if each buffer was specified previously in own code block outside
`nk_clear`.

Overall the buffer clearing was an overreaching measure of `nk_clear`
and I hope not to many tried to make use of it. In general the buffer
is owned by the user of nuklear and not nuklear itself.
Therefore it does not make sense for nuklear to access it after
`nk_convert` was successfully completed.

Sorry for any inconvenience caused by this breaking change.