fix nk_do_edit autoscroll.
How to test:
Create an edit widget.
Enter a long text into the edit widget.
Press "home" so that widget autoscrolls to start of text.
Press "end" so that widget now should jump to end of text.
Previously you had to press "end" several times to jump to the end of text.
This patch fixes that (pressing "end" once will jump to the end of text).